Refactor launcher directories
