+++ BazaarPay/src/main/java/ir/cafebazaar/bazaarpay/launcher/directPay/StartDirectPayFinalizeContract.kt
@@ -1,4 +1,4 @@
-package ir.cafebazaar.bazaarpay.directPay
+package ir.cafebazaar.bazaarpay.launcher.directPay
 
 import android.app.Activity
 import android.content.Context
